+++ examples/with-markprompt-web/src/main.ts
@@ -66,7 +66,20 @@ if (el && el instanceof HTMLElement) {
           },
         },
       ],
+      defaultView: {
+        message:
+          "Welcome to Markprompt! We're here to assist you. Just type your question to get started.",
+        promptsHeading: 'Popular questions',
+        prompts: [
+          'What is Markprompt?',
+          'How do I setup the React component?',
+          'Do you have a REST API?',
+        ],
+      },
     },
     defaultView: 'chat',
+    trigger: {
+      buttonLabel: 'Ask AI',
+    },
   } satisfies MarkpromptOptions);
 }